+++ .i18nrc.json
@@ -51,7 +51,7 @@
     "visTypeTable": "src/plugins/vis_type_table",
     "visTypeTagCloud": "src/plugins/vis_type_tagcloud",
     "visTypeTimeseries": ["src/legacy/core_plugins/vis_type_timeseries", "src/plugins/vis_type_timeseries"],
-    "visTypeVega": "src/legacy/core_plugins/vis_type_vega",
+    "visTypeVega": "src/plugins/vis_type_vega",
     "visTypeVislib": "src/legacy/core_plugins/vis_type_vislib",
     "visTypeXy": "src/legacy/core_plugins/vis_type_xy",
     "visualizations": "src/plugins/visualizations",
