+++ packages/kbn-optimizer/src/worker/webpack.config.ts
@@ -136,9 +136,12 @@ export function getWebpackConfig(bundle: Bundle, worker: WorkerConfig) {
 
     module: {
       // no parse rules for a few known large packages which have no require() statements
+      // or which have require() statements that should be ignored because the file is
+      // already bundled with all its necessary depedencies
       noParse: [
         /[\///]node_modules[\///]elasticsearch-browser[\///]/,
-        /[\///]node_modules[\///]lodash[\///]index\.js/,
+        /[\///]node_modules[\///]lodash[\///]index\.js$/,
+        /[\///]node_modules[\///]vega-lib[\///]build[\///]vega\.js$/,
       ],
 
       rules: [
