+++ src/legacy/core_plugins/kibana/public/__tests__/vis_type_vega/vega_map_image_256.png
