+++ src/legacy/core_plugins/kibana/public/__tests__/vis_type_vega/vegalite_image_512.png
