--- src/legacy/core_plugins/vis_type_vega/package.json
@@ -1,6 +1,0 @@
-{
-  "author": "Yuri Astrakhan<yuri@elastic.co>",
-  "name": "vega",
-  "version": "kibana"
-}
-
