+++ src/legacy/ui/ui_exports/ui_export_defaults.js
@@ -24,6 +24,7 @@ export const UI_EXPORT_DEFAULTS = {
   webpackNoParseRules: [
     /node_modules[\/\\](angular|elasticsearch-browser)[\/\\]/,
     /node_modules[\/\\](mocha|moment)[\/\\]/,
+    /node_modules[\/\\]vega-lib[\/\\]build[\/\\]vega\.js$/,
   ],
 
   webpackAliases: {
