+++ src/plugins/vis_type_vega/kibana.json
@@ -2,5 +2,6 @@
   "id": "visTypeVega",
   "version": "kibana",
   "server": true,
-  "ui": true
+  "ui": true,
+  "requiredPlugins": ["data", "visualizations", "mapsLegacy", "expressions"]
 }
