+++ src/plugins/vis_type_vega/public/_vega_editor.scss
