+++ src/plugins/vis_type_vega/public/components/index.ts
