+++ src/plugins/vis_type_vega/public/components/vega_actions_menu.tsx
