+++ src/plugins/vis_type_vega/public/components/vega_vis_editor.tsx
