+++ src/plugins/vis_type_vega/public/data_model/ems_file_parser.js
