+++ src/plugins/vis_type_vega/public/data_model/es_query_parser.test.js
