+++ src/plugins/vis_type_vega/public/data_model/search_cache.js
