+++ src/plugins/vis_type_vega/public/data_model/vega_parser.js
