+++ src/plugins/vis_type_vega/public/default.spec.hjson
