+++ src/plugins/vis_type_vega/public/index.scss
@@ -1,5 +1,3 @@
-@import 'src/legacy/ui/public/styles/styling_constants';
-
 // Prefix all styles with "vga" to avoid conflicts.
 // Examples
 //   vgaChart
