+++ src/plugins/vis_type_vega/public/vega_fn.ts
@@ -19,11 +19,7 @@
 
 import { get } from 'lodash';
 import { i18n } from '@kbn/i18n';
-import {
-  ExpressionFunctionDefinition,
-  KibanaContext,
-  Render,
-} from '../../../../plugins/expressions/public';
+import { ExpressionFunctionDefinition, KibanaContext, Render } from '../../expressions/public';
 import { VegaVisualizationDependencies } from './plugin';
 import { createVegaRequestHandler } from './vega_request_handler';
