+++ src/plugins/vis_type_vega/public/vega_type.ts
@@ -18,10 +18,10 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { DefaultEditorSize } from '../../../../plugins/vis_default_editor/public';
+import { DefaultEditorSize } from '../../vis_default_editor/public';
 import { VegaVisualizationDependencies } from './plugin';
 import { VegaVisEditor } from './components';
-import { defaultFeedbackMessage } from '../../../../plugins/kibana_utils/public';
+import { defaultFeedbackMessage } from '../../kibana_utils/public';
 
 import { createVegaRequestHandler } from './vega_request_handler';
 // @ts-ignore
