+++ src/plugins/vis_type_vega/public/vega_view/vega_map_layer.js
@@ -17,9 +17,26 @@
  * under the License.
  */
 
-import { PluginInitializerContext } from '../../../../core/public';
-import { VegaPlugin as Plugin } from './plugin';
+import L from 'leaflet';
+import 'leaflet-vega';
+import { KibanaMapLayer } from '../../../maps_legacy/public';
 
-export function plugin(initializerContext: PluginInitializerContext) {
-  return new Plugin(initializerContext);
+export class VegaMapLayer extends KibanaMapLayer {
+  constructor(spec, options) {
+    super();
+
+    // Used by super.getAttributions()
+    this._attribution = options.attribution;
+    delete options.attribution;
+
+    this._leafletLayer = L.vega(spec, options);
+  }
+
+  getVegaView() {
+    return this._leafletLayer._view;
+  }
+
+  getVegaSpec() {
+    return this._leafletLayer._spec;
+  }
 }
