+++ src/plugins/vis_type_vega/public/vega_view/vega_tooltip.js
