+++ src/plugins/vis_type_vega/public/vega_visualization.js
